refactor: queries/fragments on relay-compiler manner
