+++ codegen.yml
@@ -1,15 +1,23 @@
 schema: schema.graphql
 documents: 'src/**/*.graphql'
 generates:
-  src/__generated__/types.tsx:
-    config:
-      nonOptionalTypename: true
-      withComponent: true
-      withHOC: false
-      withHooks: false
-      reactApolloVersion: 2
-      dedupeOperationSuffix: true
+  src/__generated__/types.ts:
+    - typescript
+  src/:
+    preset: near-operation-file
+    presetConfig:
+      folder: __generated__
+      extension: .tsx
+      baseTypesPath: __generated__/types.ts
     plugins:
-      - typescript
       - typescript-operations
       - typescript-react-apollo
+    config:
+      namingConvention: keep
+      nonOptionalTypename: true
+      dedupeOperationSuffix: true
+      withComponent: true
+      withHooks: true
+      withHOC: false
+      reactApolloVersion: 2
+      # documentMode: documentNode
