+++ src/pages/_routes.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { Route, Switch } from 'react-router-dom';
 import Index from './';
 import Error from './_error';
-import Orders from './orders';
+import Orders from './orders/OrderQuery';
 
 export default function AppRoute() {
   return (
