+++ src/pages/customers/Customer_data.graphql
@@ -1,6 +1,6 @@
 fragment Customer_data on Customer {
   companyName
-  orderList {
-    orderID
+  address {
+    city
   }
 }
